requirements.txt
# Core dependencies
llama-cpp-python>=0.2.0
numpy>=1.24.0
pyyaml>=6.0

# Data handling and analysis
pandas>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Jupyter for examples
jupyter>=1.0.0
ipykernel>=6.25.0
